backend/pyproject.toml
 [project]
 name = "takeoff-ai-backend"
 version = "0.1.0"
 description = "API FastAPI para o MVP TAKEOFF.AI (upload DWG/DXF, parsing e relatórios)"
 readme = "README.md"
 requires-python = ">=3.10"
 authors = [
   { name = "Matheus / DELMAT Engenharia" }
 ]
 dependencies = [
   "fastapi>=0.111.0",
   "uvicorn[standard]>=0.30.0",
   "sqlmodel>=0.0.22",
   "asyncpg>=0.29.0",
   "alembic>=1.13.3",
   "python-multipart>=0.0.9",
   "pydantic-settings>=2.4.0",
   "python-jose[cryptography]>=3.3.0",
   "passlib[bcrypt]>=1.7.4",
   "bcrypt>=4.1.2",
   "httpx>=0.27.2",
   "ezdxf>=1.2.0",
   "openpyxl>=3.1.2",
   "redis>=5.0.4",
   "aiofiles>=24.1.0",
   "boto3>=1.34.70",
   "loguru>=0.7.2",
   "python-dateutil>=2.9.0",
   "pandas>=2.2.2"
 ]

 [project.optional-dependencies]
 dev = [
   "pytest>=8.1.1",
   "pytest-asyncio>=0.23.7",
   "httpx>=0.27.2",
   "ruff>=0.3.4"
 ]

 [build-system]
 requires = ["setuptools>=69.2", "wheel"]
 build-backend = "setuptools.build_meta"
